requirements.txt
requests
boto3
pillow
sanic
ydb
telegram